Consolidated databaseConverters
