+++ src/utils/index.ts
@@ -1,5 +1,4 @@
 export * from "./parsePhone";
 export * from "./saveUserData";
 export * from "./getSavedUserData";
-export * from "./userConverter";
-export * from "./jamAttendanceConverter";
+export * from "./firebaseDataConverter";
